keep健身后端/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
PyMySQL==1.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
cryptography==41.0.3

# JWT认证
PyJWT==2.8.0

# HTTP请求（第三方登录）
requests==2.31.0

# 密码加密
Werkzeug==2.3.7

# 数据验证
pydantic==2.5.0

# 异步任务处理
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# 视频处理(需要系统安装ffmpeg)
# ffmpeg-python==0.2.0

# 可选依赖
# Flask-Caching==2.0.2
# Flask-Mail==0.9.1
# Pillow==10.0.0
# python-jose==3.3.0  # 用于Apple登录签名验证
